Update brightness parameters in effect classes to use float values
- Changed baseBrightness from uint8_t to float in CodeRainEffect, RippleEffect, and ZenLightsEffect to allow for a range of 0.0 to 1.0.
- Updated preset definitions and parameter handling to reflect the new float type for baseBrightness.
- Adjusted calculations in Update methods to accommodate the new baseBrightness type.

+++ lib/CodeRainEffect/src/CodeRainEffect.h
@@ -18,7 +18,7 @@ public:
         float baseHue; // 0.0 to 1.0
         float hueVariation;
         float saturation;
-        uint8_t baseBrightness; // 0-255, for HSB conversion
+        float baseBrightness; // 0-1, for HSB conversion
     };
 
     static const Parameters ClassicMatrixPreset;
